Fix a type issue for chunk_size
chunk_size was stored as int, but the chunk_size can be > signed 32 bit. On my local machine,
a 24 K x 24 K image caused it to overflow, whereas a 16 K x 16 K image seemed to work.

Seems like this was introduced in/around commit 59c367f13 , which
integrates the phase shift in CPU degridders.

+++ idg-lib/src/common/ArrayTypes.h
@@ -119,9 +119,9 @@ class Array1D {
    * @return 'start' and 'next' indices for use in openmp blocks.
    */
   std::pair<size_t, size_t> omp_range() const {
-    const int thread_num = omp_get_thread_num();
-    const bool last = (thread_num + 1 == omp_get_num_threads());
-    const int chunk_size = size() / omp_get_num_threads();
+    const size_t thread_num = omp_get_thread_num();
+    const bool last = (thread_num + 1 == size_t(omp_get_num_threads()));
+    const size_t chunk_size = size() / omp_get_num_threads();
     size_t start = chunk_size * thread_num;
     size_t next = last ? size() : start + chunk_size;
     return std::make_pair(start, next);
